Summon rework
It turns out that splitting half of the summon action to
an event was a bad idea, because it can't be guaranteed that
the listener will be called immediately rather than after one
or more external [event] handlers.

Other than that, this commit does the following:
- Fixes OOS in MP from usage of select event.
- Implements new validation for locations, including more
sensible restrictions regarding castles.
- Makes locations more flexible (e.g. it adds support for
 summoning water elementals on water if desired.)
- Allows cancelling summon of {UNIT} through try summon {UNIT} event.

Third and fourth feature are required for To Lands Unknown.